+++ include/monolish_equation.hpp
@@ -79,7 +79,8 @@ public:
    * @param[in] A CRS format Matrix
    * @param[in] x solution vector
    * @param[in] b right hand vector
-   * @return error code, see https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
+   * @return error code, see
+   *https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
    **/
   [[nodiscard]] int solve(MATRIX &A, vector<Float> &x, vector<Float> &b);
 
@@ -128,7 +129,8 @@ public:
 template <typename MATRIX, typename Float>
 class BiCGSTAB : public monolish::solver::solver<MATRIX, Float> {
 private:
-  [[nodiscard]] int monolish_BiCGSTAB(MATRIX &A, vector<Float> &x, vector<Float> &b);
+  [[nodiscard]] int monolish_BiCGSTAB(MATRIX &A, vector<Float> &x,
+                                      vector<Float> &b);
 
 public:
   /**
@@ -136,7 +138,8 @@ public:
    * @param[in] A CRS format Matrix
    * @param[in] x solution vector
    * @param[in] b right hand vector
-   * @return error code, see https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
+   * @return error code, see
+   *https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
    **/
   [[nodiscard]] int solve(MATRIX &A, vector<Float> &x, vector<Float> &b);
 
@@ -155,7 +158,9 @@ public:
    * - Multi-threading: false
    * - GPU acceleration: false
    **/
-  [[nodiscard]] std::string name() const { return "monolish::equation::BiCGSTAB"; }
+  [[nodiscard]] std::string name() const {
+    return "monolish::equation::BiCGSTAB";
+  }
 
   /**
    * @brief get solver name "BiCGSTAB"
@@ -185,7 +190,8 @@ public:
 template <typename MATRIX, typename Float>
 class Jacobi : public monolish::solver::solver<MATRIX, Float> {
 private:
-  [[nodiscard]] int monolish_Jacobi(MATRIX &A, vector<Float> &x, vector<Float> &b);
+  [[nodiscard]] int monolish_Jacobi(MATRIX &A, vector<Float> &x,
+                                    vector<Float> &b);
 
 public:
   /**
@@ -193,7 +199,8 @@ public:
    * @param[in] A CRS format Matrix
    * @param[in] x solution vector
    * @param[in] b right hand vector
-   * @return error code, see https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
+   * @return error code, see
+   *https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
    **/
   [[nodiscard]] int solve(MATRIX &A, vector<Float> &x, vector<Float> &b);
   void create_precond(MATRIX &A);
@@ -206,7 +213,9 @@ public:
    * - Multi-threading: false
    * - GPU acceleration: false
    **/
-  [[nodiscard]] std::string name() const { return "monolish::equation::Jacobi"; }
+  [[nodiscard]] std::string name() const {
+    return "monolish::equation::Jacobi";
+  }
 
   /**
    * @brief get solver name "Jacobi"
@@ -249,7 +258,8 @@ public:
    * @param[in] A CRS format Matrix
    * @param[in] x solution vector
    * @param[in] b right hand vector
-   * @return error code, see https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
+   * @return error code, see
+   *https://github.com/ricosjp/monolish/blob/master/include/common/monolish_common.hpp
    * @warning
    * SOR is not completely parallelized.
    * The part of solving the lower triangular matrix is performed sequentially.
@@ -299,7 +309,8 @@ class LU : public monolish::solver::solver<MATRIX, Float> {
 private:
   int lib = 1; // lib is 1
   [[nodiscard]] int mumps_LU(MATRIX &A, vector<double> &x, vector<double> &b);
-  [[nodiscard]] int cusolver_LU(MATRIX &A, vector<double> &x, vector<double> &b);
+  [[nodiscard]] int cusolver_LU(MATRIX &A, vector<double> &x,
+                                vector<double> &b);
 
 public:
   [[nodiscard]] int solve(MATRIX &A, vector<Float> &x, vector<Float> &b);
@@ -350,7 +361,8 @@ template <typename MATRIX, typename Float>
 class QR : public monolish::solver::solver<MATRIX, Float> {
 private:
   int lib = 1; // lib is 1
-  [[nodiscard]] int cusolver_QR(MATRIX &A, vector<double> &x, vector<double> &b);
+  [[nodiscard]] int cusolver_QR(MATRIX &A, vector<double> &x,
+                                vector<double> &b);
   [[nodiscard]] int cusolver_QR(MATRIX &A, vector<float> &x, vector<float> &b);
 
 public:
@@ -403,8 +415,10 @@ template <typename MATRIX, typename Float>
 class Cholesky : public monolish::solver::solver<MATRIX, Float> {
 private:
   int lib = 1; // lib is 1
-  [[nodiscard]] int cusolver_Cholesky(MATRIX &A, vector<float> &x, vector<float> &b);
-  [[nodiscard]] int cusolver_Cholesky(MATRIX &A, vector<double> &x, vector<double> &b);
+  [[nodiscard]] int cusolver_Cholesky(MATRIX &A, vector<float> &x,
+                                      vector<float> &b);
+  [[nodiscard]] int cusolver_Cholesky(MATRIX &A, vector<double> &x,
+                                      vector<double> &b);
 
 public:
   /**
@@ -427,7 +441,9 @@ public:
    * - Multi-threading: false
    * - GPU acceleration: false
    **/
-  [[nodiscard]] std::string name() const { return "monolish::equation::Cholesky"; }
+  [[nodiscard]] std::string name() const {
+    return "monolish::equation::Cholesky";
+  }
 
   /**
    * @brief get solver name "Cholesky"