+++ include/monolish_generalized_eigen.hpp
@@ -30,11 +30,11 @@ class LOBPCG : public solver::solver<MATRIX, Float> {
 private:
   // TODO: support multiple lambda(eigenvalue)s
   [[nodiscad]] int monolish_LOBPCG(MATRIX &A, MATRIX &B, vector<Float> &lambda,
-                      matrix::Dense<Float> &x, int itype);
+                                   matrix::Dense<Float> &x, int itype);
 
 public:
   [[nodiscad]] int solve(MATRIX &A, MATRIX &B, vector<Float> &lambda,
-            matrix::Dense<Float> &x, int itype);
+                         matrix::Dense<Float> &x, int itype);
 
   void create_precond(MATRIX &A) {
     throw std::runtime_error("this precond. is not impl.");
@@ -51,7 +51,9 @@ public:
    * - Multi-threading: false
    * - GPU acceleration: false
    **/
-  [[nodiscad]] std::string name() const { return "monolish::generalized_eigen::LOBPCG"; }
+  [[nodiscad]] std::string name() const {
+    return "monolish::generalized_eigen::LOBPCG";
+  }
 
   /**
    * @brief get solver name "LOBPCG"
@@ -81,10 +83,12 @@ public:
 template <typename MATRIX, typename Float>
 class DC : public solver::solver<MATRIX, Float> {
 private:
-  [[nodiscad]] int LAPACK_DC(MATRIX &A, MATRIX &B, vector<Float> &lambda, int itype);
+  [[nodiscad]] int LAPACK_DC(MATRIX &A, MATRIX &B, vector<Float> &lambda,
+                             int itype);
 
 public:
-  [[nodiscad]] int solve(MATRIX &A, MATRIX &B, vector<Float> &lambda, int itype);
+  [[nodiscad]] int solve(MATRIX &A, MATRIX &B, vector<Float> &lambda,
+                         int itype);
 
   void create_precond(MATRIX &A) {
     throw std::runtime_error("this precond. is not impl.");
@@ -101,7 +105,9 @@ public:
    * - Multi-threading: false
    * - GPU acceleration: false
    **/
-  [[nodiscad]] std::string name() const { return "monolish::generalized_eigen::DC"; }
+  [[nodiscad]] std::string name() const {
+    return "monolish::generalized_eigen::DC";
+  }
 
   /**
    * @brief get solver name "DC"