+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Unreleased
 - Fix typo in documents https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/351
 
 ### Changed
+- Add nodiscard to solve() https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/400
 - Change monolish container name oss_nvidia -> oss-nvidia, mkl_nvidia -> mkl-nvidia https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/398
 - Generate monolish container document by cmake https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/399
 - Use update-alternative to switch libmonolish_gpu.so https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/396